+++ packages/core-web/src/media/metrics.ts
@@ -1,8 +1,8 @@
 import {
+  type InitialProps,
   addLegacyMediaMetricsToStore,
   addMetricsToStore,
   createControllerStore,
-  type InitialProps,
 } from "@livepeer/core/media";
 import { createStorage, noopStorage } from "@livepeer/core/storage";
 import { version } from "@livepeer/core/version";
@@ -21,6 +21,11 @@ export type MediaMetricsOptions = Pick<InitialProps, "onError" | "viewerId"> & {
    * If not specified, the function defaults to parsing the `src` attribute of the HTMLMediaElement to get the playback ID.
    */
   playbackId?: string;
+
+  /**
+   * Disables the `progress` event listener, which is used to monitor when media is in a "playing" state.
+   */
+  disableProgressListener?: boolean;
 };
 
 /**
@@ -62,9 +67,13 @@ export function addMediaMetrics(
 
     const { destroy: destroyListeners } = addEventListeners(element, store);
 
-    const { destroy: destroyMetrics } = addMetricsToStore(store);
+    const { destroy: destroyMetrics } = addMetricsToStore(store, {
+      disableProgressListener: opts.disableProgressListener,
+    });
     const { destroy: destroyLegacyMetrics, metrics: legacyMetrics } =
-      addLegacyMediaMetricsToStore(store);
+      addLegacyMediaMetricsToStore(store, {
+        disableProgressListener: opts.disableProgressListener,
+      });
 
     store
       .getState()