+++ packages/react/CHANGELOG.md
@@ -185,6 +185,39 @@
   - @livepeer/core-web@4.2.0-next.0
   - @livepeer/core@3.2.0-next.0
 
+## 4.1.13
+
+### Patch Changes
+
+- [#521](https://github.com/livepeer/ui-kit/pull/521) [`6082120`](https://github.com/livepeer/ui-kit/commit/6082120c32e2f0417fb6473637840ccda18f1225) Thanks [@0xcadams](https://github.com/0xcadams)! - **Feature:** added audio and video constraints to the Broadcast Root component.
+
+- Updated dependencies [[`6082120`](https://github.com/livepeer/ui-kit/commit/6082120c32e2f0417fb6473637840ccda18f1225)]:
+  - @livepeer/core-web@4.1.13
+  - @livepeer/core@3.1.13
+  - @livepeer/core-react@3.1.13
+
+## 4.1.12
+
+### Patch Changes
+
+- [#518](https://github.com/livepeer/ui-kit/pull/518) [`95b9592`](https://github.com/livepeer/ui-kit/commit/95b959276dd5d23b1a26baa2ba881e149492c9f7) Thanks [@0xcadams](https://github.com/0xcadams)! - **Fix:** removed `element.load()` side effect from the `addEventListeners` function.
+
+- Updated dependencies [[`95b9592`](https://github.com/livepeer/ui-kit/commit/95b959276dd5d23b1a26baa2ba881e149492c9f7)]:
+  - @livepeer/core-web@4.1.12
+  - @livepeer/core@3.1.12
+  - @livepeer/core-react@3.1.12
+
+## 4.1.11
+
+### Patch Changes
+
+- [#515](https://github.com/livepeer/ui-kit/pull/515) [`922ff14`](https://github.com/livepeer/ui-kit/commit/922ff141ce496f78843b6fc211847725d2e57d80) Thanks [@0xcadams](https://github.com/0xcadams)! - **Feature:** added `disableProgressListener` to `addMediaMetrics` so progress events from an HTML5 video element can be ignored when monitoring for playing/paused.
+
+- Updated dependencies [[`922ff14`](https://github.com/livepeer/ui-kit/commit/922ff141ce496f78843b6fc211847725d2e57d80)]:
+  - @livepeer/core-web@4.1.11
+  - @livepeer/core@3.1.11
+  - @livepeer/core-react@3.1.11
+
 ## 4.1.10
 
 ### Patch Changes